deployments/mongo: atomically return old device status on status update
refactored the existing method to return the old status in one go with the update.

this will come in handy in the model layer, when we update aggregated
deployment statistics - we need both the old and new status to do that.

the Query.Apply + Change is a counterpart of mongo's findAndModify(), which can
do the update and retrieval atomically. while the atomicity is not required as much
as e.g. with aggregate deployment stat updates, this seemed a simpler pattern than
fetch-update-modify + writing related methods and tests.

Issues: MEN-560
Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>