deployment: bugfix for missing 'status' field in Deployment
everything seems to be in place for 'status' to be returned, i.e.
the field is correctly defined and updated in the ApiDeploymentWrapper.

the culprit is the custom marshaller for Deployment - it's called when we serialize
the wrapper struct (embedding the Deployment). for some reason it does not serialize
the 'status' field; simply removing the marshaller 'fixes' the problem.

it's not clear why this happens - a poc on go playground exhibits the correct behavior
even with the marshaller present. the undesired beahvior is observed also in our
docker containers, and on local machines.

Issues: MEN-613

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>