device_deployment_log: relax 'messages' validation
the server doesn't accept empty 'messages' list since the field is 'required'
(nil/empty slice is rejected); solution: make it 'optional'.

side effect: the server will now accept requests even if the 'messages' field is missing.
this is still persisted as 'messages: []' in mongo, so in case of a mistake, we're covered.

Issues: MEN-611

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>