images/mongo: handle the artifact's 'device_types_compatible' fields
accidentally, nothing else has to be changed for multi-device-type artifact support.
mongo handles the query correctly both for single-value dev type, and an array of dev types.

regarding the index - it becomes a multikey compound index, and works correctly.
the 'unique' property makes less sense now, but I'm keeping it since we still
wouldn't want an image with the same name and exactly the same dev types as an
existing one.

Issues: MEN-721

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>