Upload handling changed.
Artifact uploaded to the deployments service will be uploaded to file
storage using presigned link and parsed by artifact library concurrently.
The solution is based on the io.TeeReader and io.Pipe.
Both operations needs to be done in the concurrent goroutines because
reading from the Pipe is a blocking operation.

Signed-off-by: Krzysztof Jaskiewicz <krzysztof.jaskiewicz@rndity.com>