deployments/deployment: get rid of the 'aborted' deployment state
it's an undocumented state which got here by accident. remove
it to be consistent with the docs, and with GET /deployments,
which doesn't accept an 'aborted' filter. this change makes 'aborted'
device deployments count towards 'finished' deployment state.

this was ok'd by Michael - 'aborted' is not used in the ui.

Issues: MEN-935

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>