refactor http status codes in DELETE /deployments/devices/{id}
all errors from the model are masked with http 500. there's
one operation (DeviceDeploymentsStorage.DecommissionDeviceDeployments),
which could legitimately return a 'not found' error. this would be a candidate
for 404, so we're making it a 204 as it's the point of this task.

this commit introduces a new predefined storage error that we can switch
on to isolate this condition. all other lower-layer errors are still treated
as 500 internal error.

Issues: MEN-1143

Changelog: None

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>