deployments/mongo: do create a short-named deployment+artifact name index
the crucial thing is to set the Name on the index - this prevents index
name autogeneration, which can easily shoot past the length limit.

EnsureIndexKey always leaves the index.Name == "", hence autogeneration
occurs.

use EnsureIndex instead, which gives access to all index fields, and
set the name to a sane value.

(the internal mechanism is nicely explained in this patch:
https://github.com/go-mgo/mgo/commit/9be224d57f036cca92520e88c32e411682385365)

Changelog: None

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>