deployments/mongo: expose/refactor ensureIndexing
expose the ensureIndex in two versions:
1. accepts a context, possibly with tenant id - for dealing with
runtime requests where middleware sets up the context properly
2. accepts the full db name directly - for migrations, where it's more convenient

version 2 was introduced because a critical bug was observed in *all*
other services' migration code. typically the Migrations expect a context
set up just like our middlewares would do it (and use ...DbFromContext).
this doesn't work because nobody sets up this context - both store.Migrate and
migrator.Apply work directly on full db names (strings).

this omission is an error, but also - setting up the context when we have a db
name ready doesn't make any sense. hence the split - Migrations will be set up
with the db name directly, and use ver 2 of the indexing method.

ps. a set of bugs will be filed for other services, this is critical for MT

Changelog: None

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>