Handle error from http.NewRequest() in GetDeviceInventory()
According to sources for NewRequest:

https://golang.org/src/net/http/request.go?s=28198:28299#L828

It can return error in case URL is bad, context is nil, or
HTTP method is bad. I guess here, the most likely failure
would be a badly formed URL.

I assume that this "never happens", because it hasn't been a
problem, but it seems more correct to handle the error, just
in case.

Reported by LGTM:
https://lgtm.com/rules/1510373626114/

Changelog: None
Ticket: None
Signed-off-by: Ole Herman Schumacher Elgesem <ole@northern.tech>