Refactor NewImage and NewImageForTenantHandler
NewImage and NewImageForTenantHandler overlap: the first is the
management API which requires the user to be authenticated via JWT
token, the latter is the internal end-point used to upload an artifact
for a given tenant without authentication.

Their implementations overlap, thus we refactor the functions to share
the common part.

changelog: none
Signed-off-by: Fabio Tranchitella <fabio@tranchitella.eu>