Avoid minio.error.InvalidEndpointError with latest Python
With latest Python releases we get the following exception:

    minio.error.InvalidEndpointError: InvalidEndpointError:
    message: Hostname cannot have a scheme.

Python 3.8's urlparse interprets `minio:9000` as schema `minio` followed
by location `9000`. This breaks our tests as the minio client library
does not support schemas in endpoint URIs.

Having '//' in the beginning of the endpoint enforces urlsplit to
consider the endpoint as a netloc according to this quote in
docs.python.org/3/library/urllib.parse.html:

   Following the syntax specifications in RFC 1808, urlparse
   recognizes a netloc only if it is properly introduced by ‘//’.
   Otherwise the input is presumed to be a relative URL and thus
   to start with a path component.

A new version of the minio client library containing the fix is now
available, and it includes the following PR:

    https://github.com/minio/minio-py/pull/836

changelog: none
Signed-off-by: Fabio Tranchitella <fabio@tranchitella.eu>