Remove the size parameter from /artifacts/generate end-point
The size parameter is not needed (anymoreand can be removed.

At the same time, we can make it optional in the other end-points
dealing with artifact upload (`POST /tenants/{id}/artifacts` and
`POST /artifacts`).

If it is provided, it is required to be equal to the size of the file
uploaded.

Changelog: none
Signed-off-by: Fabio Tranchitella <fabio@tranchitella.eu>