Install musl runtime in tester Docker image to execute production binary
Production `deployments` binary is build from an Alpine based container
(i.e. musl linker) so we need to support this from the Debian based
runner to be able to test the same binary that we would send to
production.

Note that this was previously workarounded by just rebuilding the binary
in the CI Pipeline with CGO disabled, producing an statically linked
binary but dropping LZMA support.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>