Upload image streams the content directly to s3 while processing request
Caching the artifact has the drawback of leaving the connection idle for
quite long, risking the connection to time out if the artifact is large.
This commit fixes that by parsing the form-data as a stream and
uploading the artifact as it is received.
As an added bonus: The progress bar in the UI is actually representative
of upload progress.

The size parameter of POST /deployments/artifacts is no longer required.
Instead the size is determined from the uploaded size (the max limit
stil applies).

changelog: none

Signed-off-by: Alf-Rune Siqveland <alf.rune@northern.tech>