store: sanitize statistics update functions
from now on these only update stats, and are
renamed to reflect how exactly.

UpdateStatsInc incrementally incs/decs statistics
counters by device status.

UpdateStats just updates the stats wholesale, as passed
to the func.

do not update the deployment status inside any of them, because
it's a confusing workaround and it's hard to track what
changes where.

db funcs are now dumb and specialized, and will be glued together
in the upper layer.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>