Fix migration 1.2.3
The migration fails to query the images (artifacts) in multi-tenant
environments, leading to index creation error due to the fact the
documents are not migrated to the new depends and depends_idx
structures.

Instead of relying on the store, use the mongo driver directly to
query the collection in the right (tenant-wise) database.

Changelog: none

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>