config: Add enrich configuration options for signed URLs
I added options for fallback values in case the proxy is not configured
to forward X-Forwarded-Host header as well as URL scheme. A provisional
signature algorithm parmeter is also added in case there will be a need
for changing the algorithm (though it's lacking implementation).

The application also issues a warning if the presign secret has not been
altered from the default value.

changelog: none

Signed-off-by: Alf-Rune Siqveland <alf.rune@northern.tech>