Remove the gzip middleware
We delegate the response compression to the API gateway. For this
reason, we don't want to enable the gzip middleware directly in the
service. The Content-Encoding: gzip causes the side effect of response
chunking by the API gateway (Traefik), because it needs to decompress
and re-compress the payload. The chunked transfer breaks the Mender
client's artifact download, which expects a Content-Length header.

Changelog: none

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>