Increase the S3 upload timeout to 60 minutes and make it configurable
The Mender API gateway, based on Traefik, doesn't buffer the file
uploads. This slows down the upload to S3 to match the upload speed of
the client; in turn, this requires an adjustement of the expiration time
of the S3 presigned upload URL to allow higher upload times.

Both upload and download links' expiry duration are now configurable:

* aws.download_expire_seconds (defaults to 900 seconds)
* aws.upload_expire_seconds (defaults to 3600 seconds)

Changelog: none

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>