tests: capture (deprecated) pytest.config manually
pytest.config used to be available globally, and before
test collection. now it's deprecated/removed (since pytest 5.x).

there are a couple ways to capture it. I wanted a drop-in replacement
for existing tests so now an alternative global var is defined in
a new 'config' module, initialized at the 'pytest_configure' hook.
(conftest would be a better place but importing from it is discouraged).

other options were:
- wrapping all clients in fixtures (request.config is now available),
but we create them manually in var. places, too much refactoring of test func
signatures etc.
- global session/autouse fixture available without explicit usage - couldn't get
it right.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>