tests/client: use config.pytest_config
the old pytest.config was available at test collection/import time,
so it was ok for initializing class vars. this is mostly where it was
used.

config.pytest_config will be used now, but it's not available that
early in the cycle. we have to move initializations to a later time,
the most convenient is at instance __init__s.

adjust the inheritance mechanism accordingly (super().__init__s, etc.).

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>