MEN-4525: move the /deployments/devices/{id} end-point to internal
We move the decommissioning end-point from management to internal, where
it really belongs to. This commit breaks back-compatibility, but the
endpoint wasn't meant to be used by users and it is safer to remove it
straight away.

This change deserves a bump in the major version of deployments.

Changelog: title

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>