Change alive and healthcheck status code to 200
Currently these two APIs return status code 204 (no content), which is
fine but it doesn't play well with systems that require status code 200
for healthcheck and alive APIs. One of these is Kubernetes Ingress. The
Ingress by default (and cannot be changed) expects the status for
HTTP(s) healthcheck API be 200. This commit allows the deployment
service be configured behind an Ingress.

Signed-off-by: Amin Hassani <ahassani@nuro.ai>