refac: Use a single source of truth for artifact size
This commit moves the configuration for the maximum image size to the
`api/http` package instead of spreading it accross multiple packages. It
also changes the utils.LimitedReader to be able to return the exact
count of the bytes uploaded - as a side-effect, this commit fixes a bug
where multipart uploads always result in the stored image size to be the
maximum image size.

Changelog: fix: use correct artifact size for multipart uploads
Ticket: None
Signed-off-by: Alf-Rune Siqveland <alf.rune@northern.tech>