test: set release notes acceptace test.
The test is using requests, due to the fact that the swagger client
we wrote is not prepared for the _v2 API and a separate spec file.
I did try to introduce the second spec API docs file but there is
a number of things to change, and we are supposed to move the specs
to open api anyways.

Also:
* locking of the database, since we are using one and same db
  for all the tests
* do not drop the migration_info collection, since recent changes
  in the deployments relay on the existence of the version information

Ticket: MEN-6593
Signed-off-by: Peter Grzybowski <peter@northern.tech>