chore: back-porting helper function and parameters to NewDeployments
two backports in order to reduce number of conflicts:

* while bringing changes to ent fork extra function was needed,
due to the fact that handleArtifact crossed the QA-required complexity
of (20). back porting saveUpdateTypes to reduce the number of conflicts
in the future os->ent syncs

* the ent-os discrepancy of NewDeployments arguments is something that is
a conflicts producer and waste of time: you always have to fix the units
on the way from os to ent.

Ticket: None
Signed-off-by: Peter Grzybowski <peter@northern.tech>