perf: Optimize database interactions when updating deployment stats
Changed the sequential find and double update ops with a findAndModify
followed by a conditional update if the status actually changed.
This also improves atomicity as there is an inherent race condition when
doing the deployment status aggregation locally.

I also noticed that most internal calls to UpdateDeviceDeploymentStatus
already has the DeviceDeployment already, so I added an internal
updateDeviceDeploymentStatus which skips fetching the device deployment
(again).

Signed-off-by: Alf-Rune Siqveland <alf.rune@northern.tech>